create order entity

+++ src/main/java/org/suehay/microservicesproducts/model/entities/ProductEntity.java
@@ -21,7 +21,7 @@ public class ProductEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    @JdbcTypeCode(SqlTypes.INTEGER)
+    @JdbcTypeCode(SqlTypes.BIGINT)
     private Long id;
 
     private String name;
